ledger: remove public CheckoutCreateView API, replace with internal session tools
