+++ ledger/checkout/utils.py
@@ -1,4 +1,91 @@
+from decimal import Decimal as D, ROUND_HALF_DOWN
+
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import AnonymousUser
+from django.conf import settings
 from oscar.apps.checkout.utils import CheckoutSessionData as CoreCheckoutSessionData
+from oscar.apps.voucher.models import Voucher
+from oscar.core.loading import get_class
+from ledger.accounts.models import EmailUser
+from ledger.checkout import serializers
+from ledger.catalogue.models import Product
+from ledger.basket.models import Basket
+from ledger.basket.middleware import BasketMiddleware
+
+Selector = get_class('partner.strategy', 'Selector')
+selector = Selector()
+
+def create_basket_session(request, parameters):
+    serializer = serializers.BasketSerializer(data=parameters)
+    serializer.is_valid(raise_exception=True)
+
+    custom = serializer.validated_data.get('custom_basket')
+    # validate product list
+    if custom:
+        product_serializer = serializers.CheckoutCustomProductSerializer(data=serializer.initial_data.get('products'), many=True)
+    else:
+        product_serializer = serializers.CheckoutProductSerializer(data=serializer.initial_data.get('products'), many=True)
+    product_serializer.is_valid(raise_exception=True)
+        
+    # validate basket
+    if serializer.validated_data.get('vouchers'):
+        if custom:
+            basket = createCustomBasket(serializer.validated_data['products'],
+                                        request.user, serializer.validated_data['system'],
+                                        vouchers=serializer.validated_data['vouchers'])
+        else:
+            basket = createBasket(serializer.validated_data['products'], request.user,
+                                    serializer.validated_data['system'],
+                                    vouchers=serializer.validated_data['vouchers'])
+    else:
+        if custom:
+            basket = createCustomBasket(serializer.validated_data['products'],
+                                        request.user, serializer.validated_data['system'])
+        else:
+            basket = createBasket(serializer.validated_data['products'],
+                                    request.user, serializer.validated_data['system'])
+
+    return basket, BasketMiddleware().get_basket_hash(basket.id)
+
+
+def create_checkout_session(request, parameters):
+    serializer = serializers.CheckoutSerializer(data=parameters)
+    serializer.is_valid(raise_exception=True)
+    # fields = [
+    #     'card_method', 'basket_owner', 'template', 'fallback_url', 'return_url', 
+    #     'return_preload_url', 'associateInvoiceWithToken', 'forceRedirect', 'sendEmail', 'proxy',
+    #     'checkoutWithToken', 'bpay_format', 'icrn_format', 'invoice_text', 'check_url'
+    # ]
+
+    session_data = CheckoutSessionData(request) 
+    
+    session_data.charge_by(serializer.validated_data['card_method'])
+    session_data.owned_by(serializer.validated_data['basket_owner'])
+    # FIXME: replace internal user ID with email address once lookup/alias issues sorted
+    email = None
+    if serializer.validated_data['basket_owner'] and request.user.is_anonymous():
+        email = EmailUser.objects.get(id=serializer.validated_data['basket_owner']).email
+    session_data.set_guest_email(email)
+
+    session_data.use_template(serializer.validated_data['template'])
+
+    # fallback url?
+    session_data.return_to(serializer.validated_data['return_url'])
+    session_data.return_preload_to(serializer.validated_data['return_preload_url'])
+    session_data.associate_invoice(serializer.validated_data['associateInvoiceWithToken'])
+    session_data.redirect_forcefully(serializer.validated_data['forceRedirect'])
+    session_data.return_email(serializer.validated_data['sendEmail'])
+    session_data.is_proxy(serializer.validated_data['proxy'])
+    session_data.checkout_using_token(serializer.validated_data['checkoutWithToken'])
+
+    session_data.bpay_using(serializer.validated_data['bpay_format'])
+    session_data.icrn_using(serializer.validated_data['icrn_format'])
+    session_data.bpay_by(serializer.validated_data['icrn_date'])
+
+    session_data.set_invoice_text(serializer.validated_data['invoice_text'])
+
+    session_data.set_last_check(serializer.validated_data['check_url'])
+    
 
 class CheckoutSessionData(CoreCheckoutSessionData):
     # Custom Ledger methods
@@ -144,3 +231,134 @@ class CheckoutSessionData(CoreCheckoutSessionData):
 
     def get_last_check(self):
         return self._get('ledger','last_check')
+
+
+def calculate_excl_gst(amount):
+    result = D(100.0)/ D(100 + settings.LEDGER_GST) * D(amount)
+    return result.quantize(D('0.01'), ROUND_HALF_DOWN)
+
+
+def createBasket(product_list, owner, system, vouchers=None, force_flush=True):
+    ''' Create a basket so that a user can check it out.
+        @param product_list - [
+            {
+                "id": "<id of the product in oscar>",
+                "quantity": "<quantity of the products to be added>"
+            }
+        ]
+        @param - owner (user id or user object)
+    '''
+    try:
+        old_basket = basket = None
+        valid_products = []
+        User = get_user_model()
+        # Check if owner is of class AUTH_USER_MODEL or id
+        if not isinstance(owner, AnonymousUser):
+            if not isinstance(owner, User):
+                owner = User.objects.get(id=owner)
+            # Check if owner has previous baskets
+            if owner.baskets.filter(status='Open'):
+                old_basket = owner.baskets.get(status='Open')
+
+        # Use the previously open basket if its present or create a new one
+        if old_basket:
+            if system.lower() == old_basket.system.lower() or not old_basket.system:
+                basket = old_basket
+                if force_flush:
+                    basket.flush()
+            else:
+                raise ValidationError('You have a basket that is not completed in system {}'.format(old_basket.system))
+        else:
+            basket = Basket()
+        # Set the owner and strategy being used to create the basket
+        if isinstance(owner, User):
+            basket.owner = owner
+        basket.system = system
+        basket.strategy = selector.strategy(user=owner)
+        # Check if there are products to be added to the cart and if they are valid products
+        if not product_list:
+            raise ValueError('There are no products to add to the order.')
+        for product in product_list:
+            p = Product.objects.get(id=product["id"])
+            if not product.get("quantity"):
+                product["quantity"] = 1
+            valid_products.append({'product': p, 'quantity': product["quantity"]})
+        # Add the valid products to the basket
+        for p in valid_products:
+            basket.add_product(p['product'],p['quantity'])
+        # Add vouchers to the basket
+        if vouchers is not None:
+            for v in vouchers:
+                basket.vouchers.add(Voucher.objects.get(code=v["code"]))
+        # Save the basket
+        basket.save()
+        return basket
+    except Product.DoesNotExist:
+        raise
+    except Exception as e:
+        raise
+
+
+def createCustomBasket(product_list, owner, system,vouchers=None, force_flush=True):
+    ''' Create a basket so that a user can check it out.
+        @param product_list - [
+            {
+                "id": "<id of the product in oscar>",
+                "quantity": "<quantity of the products to be added>"
+            }
+        ]
+        @param - owner (user id or user object)
+    '''
+    #import pdb; pdb.set_trace()
+    try:
+        old_basket = basket = None
+        valid_products = []
+        User = get_user_model()
+        # Check if owner is of class AUTH_USER_MODEL or id
+        if not isinstance(owner, AnonymousUser):
+            if not isinstance(owner, User):
+                owner = User.objects.get(id=owner)
+            # Check if owner has previous baskets
+            if owner.baskets.filter(status='Open'):
+                old_basket = owner.baskets.get(status='Open')
+
+        # Use the previously open basket if its present or create a new one
+        if old_basket:
+            if system.lower() == old_basket.system.lower() or not old_basket.system:
+                basket = old_basket
+                if force_flush:
+                    basket.flush()
+            else:
+                raise ValidationError('You have a basket that is not completed in system {}'.format(old_basket.system))
+        else:
+            basket = Basket()
+        # Set the owner and strategy being used to create the basket
+        if isinstance(owner, User):
+            basket.owner = owner
+        basket.system = system
+        basket.strategy = selector.strategy(user=owner)
+        basket.custom_ledger = True
+        # Check if there are products to be added to the cart and if they are valid products
+        defaults = ('ledger_description','quantity','price_incl_tax','oracle_code')
+        for p in product_list:
+            if not all(d in p for d in defaults):
+                raise ValidationError('Please make sure that the product format is valid')
+            p['price_excl_tax'] = calculate_excl_gst(p['price_incl_tax'])
+        # Save the basket
+        basket.save()
+        # Add the valid products to the basket
+        for p in product_list:
+            basket.addNonOscarProduct(p)
+        # Save the basket (again)
+        basket.save()
+        # Add vouchers to the basket
+        if vouchers is not None:
+            for v in vouchers:
+                basket.vouchers.add(Voucher.objects.get(code=v["code"]))
+            basket.save()
+        return basket
+    except Product.DoesNotExist:
+        raise
+    except Exception as e:
+        raise
+
