+++ ledger/payments/helpers.py
@@ -1,3 +1,15 @@
+from django.conf import settings
+
+
+def is_valid_system(system_id):
+    ''' Check if the system is in the itsystems register.
+    :return: Boolean
+    '''
+    if settings.VALID_SYSTEMS:
+        return system_id in settings.VALID_SYSTEMS
+    else:
+        logger.warn('VALID_SYSTEMS not set, ledger.payments.helpers.is_valid_system will always return true')
+        return True
 
 
 def belongs_to(user, group_name):
