+++ ledger/payments/urls.py
@@ -9,7 +9,6 @@ from ledger.payments.api import (
                 CashViewSet,
                 BpayFileList,
                 ReportCreateView,
-                CheckoutCreateView,
                 RegionViewSet
                 )
 
@@ -31,7 +30,6 @@ router.register(r'^regions', RegionViewSet)
 api_patterns = [
     url(r'api/bpoint/payment$', BpointPaymentCreateView.as_view(), name='bpoint-payment'),
     url(r'api/report$', ReportCreateView.as_view(),name='ledger-report'),
-    url(r'api/checkout$',CheckoutCreateView.as_view(), name='ledger-initial-checkout'),
     url(r'api/', include(router.urls)),
 ]
